refactor: move new functions to their own module
This is tidier and allows keeping both modules more cohesive.
